week4-postgresql-migration/requirements.txt
# Week 4 PostgreSQL Migration - Dependencies

# Core FastAPI and web framework
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5

# PostgreSQL and PostGIS
psycopg2-binary>=2.9.0
SQLAlchemy>=1.4.0
alembic>=1.7.0

# Spatial and geographic libraries
Shapely>=1.8.0
geopandas>=0.12.0
pyproj>=3.3.0

# Data processing
pandas>=1.3.0
numpy>=1.21.0

# HTTP requests and API testing
requests>=2.26.0
httpx>=0.24.0

# Utilities
python-dotenv>=0.19.0
pydantic>=1.8.0
typing-extensions>=4.0.0

# Development and testing
pytest>=6.2.0
pytest-asyncio>=0.15.0
pytest-cov>=3.0.0

# Code quality
black>=21.9.0
flake8>=4.0.0
isort>=5.9.0
mypy>=0.910

# Monitoring and logging
structlog>=21.1.0
rich>=10.0.0

# Performance and profiling
memory-profiler>=0.60.0
line-profiler>=3.3.0

# Security
cryptography>=3.4.0
python-jose[cryptography]>=3.3.0

# Optional: Jupyter for data analysis
jupyter>=1.0.0
ipykernel>=6.4.0

# Documentation
mkdocs>=1.2.0
mkdocs-material>=7.3.0